Revert "Add check for forbidden characters."
This reverts commit acf1d41c07f5c7ba0b90e195767a99a345a67bc8.

It turns out this is too strict, as files for other states (and
potentially CA in the future) contain columns that may have questions in
them (e.g., what is being voted on), and country/precinct names may
contain '#' and other characters, etc.

In the future, we may want to consider a separate test to warn us about
certain characters appearing, but that test should be lower priority
than the main format tests.